Add docs and code restructure.
Add docs for `PoolConfigDTO` in file `pool_config_dto.py`.
The abstract interfaces `SingleConnectionInterface` and `PoolConnectionInterface` have been combined in a single `connection_interface` file.

Major changes:
- New `connection_interface.py` (./project_code/abstract/database/);
- Delete `pool_connection_interface.py` (./project_code/abstract/database/);
- Delete `single_connection_interface.py` (./project_code/abstract/database/);

Minor changes:
- edit `sql_api_interface.py` 0.2.0 -> 0.2.1 (./project_code/abstract/api/);
- edit `test_pool_connection_interface.py` 0.2.1 -> 0.2.2 (./project_code/tests/test_abstract/);
- edit `test_single_connection_interface.py` 0.2.1 -> 0.2.2 (./project_code/tests/test_abstract/);
