+++ project_code/abstract/database/connection_interface.py
@@ -9,12 +9,25 @@ Apache license, version 2.0 (Apache-2.0 license)
 
 __all__: list[str] = [
     'PoolConnectionInterface',
+    'SingleConnectionInterface',
 ]
 
 __author__ = "4-proxy"
 __version__ = "0.1.0"
 
 from abc import ABC, abstractmethod
+
+
+# ______________________________________________________________________________________________________________________
+class SingleConnectionInterface[ConnectionType](ABC):
+    @abstractmethod
+    def create_connection_with_database(self, **dbconfig) -> ConnectionType:
+        pass
+
+    # ------------------------------------------------------------------------------------------------------------------
+    @abstractmethod
+    def get_connection_with_database(self) -> ConnectionType:
+        pass
 
 
 # ______________________________________________________________________________________________________________________
